requirements.txt
annotated-types==0.6.0
anthropic==0.25.7
anyio==3.7.1
boto3==1.34.96
botocore==1.34.96
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
defusedxml==0.7.1
distro==1.9.0
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.105.0
filelock==3.14.0
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
huggingface-hub==0.23.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
langchain-anthropic==0.1.11
langchain-aws==0.1.2
langchain-core==0.1.49
langchain-mongodb==0.1.3
langchain-openai==0.1.6
langsmith==0.1.53
MarkupSafe==2.1.3
motor==3.4.0
numpy==1.26.4
openai==1.25.1
orjson==3.10.2
packaging==23.2
pydantic==2.5.2
pydantic-extra-types==2.2.0
pydantic-settings==2.1.0
pydantic_core==2.14.5
pymongo==4.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
regex==2024.4.28
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tenacity==8.2.3
tiktoken==0.6.0
tokenizers==0.19.1
tqdm==4.66.2
types-requests==2.31.0.20240406
typing_extensions==4.9.0
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
